+++ CHANGELOG.md
@@ -2,7 +2,12 @@ PowerModelsSecurityConstrained.jl Change Log
 ============================================
 
 ### Staged
+- nothing
+
+### v0.1.1
 - Added missing `gen_default` dictionary for security-stage solvers
+- Added tests for infeasible network models
+- Improved the resilience of `goc_challenge1_huristic` to infeasible models
 - Fixed contingency area indexing bug
 
 ### v0.1.0